+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - `Innmind\Server\Status\Server::cpu()` now returns `Innmind\Immutable\Attempt<Innmind\Server\Status\Server\Cpu>`
 - `Innmind\Server\Status\Server::memory()` now returns `Innmind\Immutable\Attempt<Innmind\Server\Status\Server\Memory>`
 
+### Removed
+
+- `Innmind\Server\Status\Exception\EmptyPathNotAllowed`
+
 ## 4.1.1 - 2024-09-30
 
 ### Fixed